@@ -9,9 +9,18 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "react": "^18",
+    "@radix-ui/react-avatar": "^1.1.0",
+    "@radix-ui/react-dropdown-menu": "^2.1.1",
+    "@radix-ui/react-icons": "^1.3.0",
+    "@radix-ui/react-slot": "^1.1.0",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.1",
+    "lucide-react": "^0.404.0",
+    "next": "14.2.4",
+    "react": "^18.3.1",
     "react-dom": "^18",
-    "next": "14.2.4"
+    "tailwind-merge": "^2.4.0",
+    "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
     "postcss": "^8",
