@@ -1,17 +1,21 @@
-import { Inter } from "next/font/google";
-import "./globals.css";
+// This is the root layout component for your Next.js app.
+// Learn more: https://nextjs.org/docs/app/building-your-application/routing/pages-and-layouts#root-layout-required
 
-const inter = Inter({ subsets: ["latin"] });
+import { Inter } from 'next/font/google'
+import './styles.css'
 
-export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
+const inter = Inter({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-inter',
+})
 
-export default function RootLayout({ children }) {
+export default function Layout({ children }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.variable}>
+        {children}
+      </body>
     </html>
-  );
+  )
 }
